requirements.txt
bottle
pyyaml